PAPP-34959 | DRY refactor
